# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KODEGIRI",
        "version": "1.0.0"
    },
    "servers": [
        {
          "url": "http://localhost:9778"
        }
      ],
      "components": {
        "securitySchemes": {
          "bearerAuth": {
            "type": "http",           
            "scheme": "bearer"
          
          }
        }
      },
      
    "tags": [
        {
            "name": "EVENT"
        },
        {
            "name": "USER"
        }             
    ],
    "paths": {
        "/api/v1/event":{
            "post": {
                "tags":[
                    "EVENT"
                ],
                "summary": "POST data EVENT",
                "consumes": "multipart/form-data",
                "security": [
                    {
                      "bearerAuth": []
                    }
                  ],
                
                "requestBody": {
                    "content": {
                        "multipart/form-data":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name":{
                                        "type": "string",
                                        "format": "string"
                                    },
                                    "start_time": {
                                        "type":"date-time"
                                    },
                                    "location":{
                                        "type":"string"
                                    },
                                    "is_active":{
                                        "type":"boolean"
                                    },
                                    "image":{
                                        "type":"string",
                                        "format":"binary"
                                    }
                                }
            
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uploaded"
                    },
                    "500": {
                        "description": "Failed !!"
                    }
                }
                
            },
            "get": {
                "tags":[
                    "EVENT"
                ],
                "summary": "GET data EVENT",
                "parameters": [
                    {
                      "name": "start",
                      "in": "query"
                    },
                    {
                        "name": "limit",
                        "in": "query"
                    }
                      
                ],                
               
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Failed !!"
                    }
                }
                
            }           
    },
    "/api/v1/event/{id}":{
        "put": {
            "tags":[
                "EVENT"
            ],
            "summary": "PUT data EVENT",
            "consumes": "multipart/form-data",
            "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "type":"integer",
                  "required":"true"
                }
            ],
            "security": [
                {
                  "bearerAuth": []
                }
              ],
            
            "requestBody": {
                "content": {
                    "multipart/form-data":{
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name":{
                                    "type": "string",
                                    "format": "string"
                                },
                                "start_time": {
                                    "type":"date-time"
                                },
                                "location":{
                                    "type":"string"
                                },
                                "is_active":{
                                    "type":"boolean"
                                },
                                "image":{
                                    "type":"string",
                                    "format":"binary"
                                }
                            }
        
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Uploaded"
                },
                "500": {
                    "description": "Failed !!"
                }
            }
            
        },
        "delete": {
            "tags":[
                "EVENT"
            ],
            "summary": "DELETE EVENT",
            "consumes": "application/json",
            "produces": ["application/json"],
            "parameters": [
                {
                  "name": "id",
                  "in": "path",
                  "type":"integer",
                  "required":"true"
                }
            ],
            "security": [
                {
                  "bearerAuth": []
                }
              ],
              "requestBody": {
                "content": {
                    "application/json":{
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type":"integer"
                                }
                            }
        
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Success"
                },
                "500": {
                    "description": "Failed !!"
                }
            }
        }
},
"/api/v1/event/search":{
    "post":{
        "tags":[
            "EVENT"
        ],
        "summary": "SEARCH EVENT WHERE NAME OR LOCATION",
        "parameters": [
            {
              "name": "start",
              "in": "query"
            },
            {
                "name": "limit",
                "in": "query"
            }
              
        ],          
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
            "content": {
                "application/json":{
                    "schema": {
                        "type": "object",
                        "properties": {
                            "search": {
                                "type":"string",
                                "description": "search by name or location"
                            },
                            "sort": {
                                "type":"string",                                
                                "description": "sorting by ASC OR DESC"
                            },
                            "indicator": {
                                "type":"string",
                                "description": "indicator for sorting column name,location,start_time,created_at"
                            }
                        }
    
                    }
                }
            }
        },
        "responses": {
            "200": {
                "description": "Success"
            },
            "500": {
                "description": "Failed !!"
            }
        }
    }
},

            
    "/api/v1/auth/login":{
        "post":{
            "tags":[
                "USER"
            ],
            "summary": "Login",
            "consumes": ["application/json"],
            "produces": ["application/json"],
            "requestBody": {
                "content": {
                    "application/json":{
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type":"string"
                                },
                                "password": {
                                    "type":"string"
                                }
                            }
        
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Success"
                },
                "500": {
                    "description": "Failed !!"
                }
            }
        }
    },
    "/api/v1/auth/register":{
        "post":{
            "tags":[
                "USER"
            ],
            "summary": "Register",
            "consumes": ["application/json"],
            "produces": ["application/json"],
            "requestBody": {
                "content": {
                    "application/json":{
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type":"string"
                                },
                                "password": {
                                    "type":"string"
                                },
                                "confirm_password": {
                                    "type":"string"
                                },
                                "gender": {
                                    "type":"string"
                                },
                                "dob": {
                                    "type":"string"
                                },
                                "phone": {
                                    "type":"string"
                                },
                                "name": {
                                    "type":"string"
                                }
                            }
        
                        }
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Success"
                },
                "500": {
                    "description": "Failed !!"
                }
            }
        }
    }
    }
}